OBPIH-7373 Fixes after review
