+++ grails-app/services/org/pih/warehouse/report/ReportService.groovy
@@ -31,7 +31,7 @@ import org.pih.warehouse.forecasting.ForecastingService
 import org.pih.warehouse.core.Tag
 import org.pih.warehouse.core.UserService
 import org.pih.warehouse.data.DataService
-import org.pih.warehouse.inventory.ProductCount
+import org.pih.warehouse.inventory.InventoryCount
 import org.pih.warehouse.inventory.CycleCountItem
 import org.pih.warehouse.inventory.CycleCountService
 import org.pih.warehouse.inventory.Inventory
@@ -1177,9 +1177,8 @@ class ReportService implements ApplicationContextAware {
             Integer quantityAdjusted = it[3]
             Integer countAdjustments = it[4]
 
-            // Retrieve all product inventories completed during the given date range
-            List<TransactionType> inventoryTypes = TransactionType.findAllByTransactionCode(TransactionCode.PRODUCT_INVENTORY)
-            Long counts = ProductCount.createCriteria().get {
+            // Retrieve the number of counts during the given date range
+            Long inventoryCounts = InventoryCount.createCriteria().get {
                 projections {
                     rowCount()
                 }
@@ -1215,7 +1214,7 @@ class ReportService implements ApplicationContextAware {
                     facility: facility,
                     product: product,
                     countAdjustments: countAdjustments,
-                    countCycleCounts: counts,
+                    countCycleCounts: inventoryCounts,
                     lastCounted: lastCounted,
                     quantityAdjusted: quantityAdjusted,
                     amountAdjusted: amountAdjusted,
