+++ src/main/groovy/org/pih/warehouse/inventory/InventoryCountTypeCode.groovy
@@ -1,0 +1,12 @@
+package org.pih.warehouse.inventory
+
+enum InventoryCountTypeCode {
+    BASELINE_ADJUSTMENT,
+    ADJUSTMENT,
+    BASELINE
+
+    @Override
+    String toString() {
+        return name()
+    }
+}
